@@ -27,56 +42,43 @@
     "@sakitam-gis/affine": "^1.1.0",
     "@sakitam-gis/mercantile": "^1.0.0",
     "@mapbox/mbtiles": "^0.12.1",
-    "ali-oss": "^6.17.1",
-    "fs-extra": "^10.1.0",
-    "gdal-async": "^3.5.1",
-    "gdal": "^0.11.1",
+    "ali-oss": "^6.20.0",
+    "fs-extra": "^11.2.0",
+    "gdal-async": "^3.9.0",
     "ndarray": "^1.0.19",
     "cwise": "^1.0.10",
     "lodash": "^4.17.21",
-    "ndarray-gdal": "^1.3.0",
+    "ndarray-gdal": "^1.5.0",
     "ndarray-concat-rows": "^1.0.1",
     "ndarray-concat-cols": "^1.0.0",
     "tapable": "^2.2.1",
-    "pino": "^8.4.0"
+    "pino": "^9.3.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.10",
-    "@babel/preset-env": "^7.18.10",
-    "@babel/eslint-parser": "^7.18.9",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-decorators": "^7.18.10",
-    "@babel/plugin-proposal-private-methods": "^7.18.6",
-    "@babel/plugin-proposal-private-property-in-object": "^7.18.6",
-    "@semantic-release/commit-analyzer": "^9.0.2",
+    "@aurorafe/eslint-config-base": "^1.1.0",
+    "@aurorafe/eslint-config-ts": "^1.1.0",
+    "@aurorafe/eslint-config-prettier": "^1.1.0",
+    "@semantic-release/commit-analyzer": "^13.0.0",
     "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^8.0.5",
-    "@semantic-release/npm": "^9.0.1",
-    "@semantic-release/release-notes-generator": "^10.0.3",
-    "@types/node": "^18.6.4",
-    "@types/lodash": "^4.14.182",
-    "@typescript-eslint/eslint-plugin": "^5.32.0",
-    "@typescript-eslint/parser": "^5.32.0",
-    "@types/ndarray": "^1.0.11",
-    "@stdlib/types": "^0.0.14",
+    "@semantic-release/github": "^10.1.3",
+    "@semantic-release/npm": "^12.0.1",
+    "@semantic-release/release-notes-generator": "^14.0.1",
+    "@types/node": "^18.19.43",
+    "@types/lodash": "^4.17.7",
+    "@types/ndarray": "^1.0.14",
+    "@stdlib/types": "^0.3.2",
     "@vitest/coverage-c8": "^0.22.1",
-    "dotenv": "^16.0.1",
-    "eslint": "^8.21.0",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-promise": "^6.0.0",
-    "prettier": "^2.7.0",
-    "semantic-release": "^19.0.3",
-    "tsup": "^6.2.1",
-    "typescript": "^4.7.4",
-    "vitest": "^0.22.1",
-    "wait-for-expect": "3.0.2"
+    "dotenv": "^16.4.5",
+    "eslint": "^8.57.0",
+    "prettier": "^3.3.3",
+    "semantic-release": "^24.0.0",
+    "tsx": "^4.16.5",
+    "tsup": "^8.2.4",
+    "typescript": "^5.3.5",
+    "vitest": "^2.0.5"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=18"
   },
   "publishConfig": {
     "access": "public"
